requirements.txt
joblib==1.3.1
matplotlib==3.7.2
numpy==1.23.0
pandas==2.2.1
Pillow==10.0.0
scikit-learn==1.3.0
seaborn==0.12.2
streamlit==1.32.2
pymongo==4.6.2
